+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `Video` component.
 - Feature: Add `width` and `height` props to `HorizontalStroke`.
 - Feature: Add `HorizontalStroke` size CSS: `TINY`, `DEFAULT`, `BIG` and `HUGE`.
+- Feature: Export `FLUID` css from `Button`.
 
 ## [1.18.1] - 2019-02-28